crypto: integrate std sha3 implementation that allows purego usage, shed off golang.org/x/crypto/sha3 usage utils: add endian file to prevent usage of x/sys/cpu preventing purego on tinygo utils: add multi-versioned file for purego or json/v2 implementations, additional to go-json one client/rpc/daemon: Have Extra be []int to allow decoding

+++ monero/client/rpc/daemon/types.go
@@ -820,7 +820,7 @@ type TransactionJSON struct {
 		} `json:"key"`
 	} `json:"vin"`
 	Vout          []TransactionOut `json:"vout"`
-	Extra         []byte           `json:"extra"`
+	Extra         []int            `json:"extra"`
 	Signatures    []types.Bytes    `json:"signatures"`
 	RctSignatures struct {
 		Type     int    `json:"type"`
